+++ backend/src/resources/user/router.ts
@@ -1,0 +1,8 @@
+import express from 'express';
+
+import UserController from './controller';
+
+const router = express.Router();
+
+router.post('/changeColor', UserController.changeColor);
+export default router;
