+++ backend/src/resources/user/service.ts
@@ -1,0 +1,9 @@
+import User from '../../models/User';
+
+const changeColor = async (id: string, color: string) => {
+  await User.update({ id }, { color });
+};
+
+export default {
+  changeColor,
+};
