+++ backend/src/server.ts
@@ -8,6 +8,7 @@ import morgan from 'morgan';
 import io from './io';
 import accRouter from './resources/account/router';
 import playListRouter from './resources/playList/router';
+import userRouter from './resources/user/router';
 
 const app = express();
 
@@ -20,6 +21,7 @@ app.use(morgan('common'));
 app.use('/assets', express.static('assets'));
 app.use('/playList', playListRouter);
 app.use('/', accRouter);
+app.use('/user', userRouter);
 app.get('/', (req, res) => {
   res.send('Hello World!');
 });
