+++ frontend/src/api/user.ts
@@ -1,13 +1,15 @@
 import { BACKEND_URL, HEADERS as headers } from '~/constants/index';
 
 export const changeColor = async (id: string, color: string) => {
-  return await fetch(`${BACKEND_URL}/changeColor`, {
-    method: 'POST',
-    headers,
-    credentials: 'include',
-    body: JSON.stringify({
-      id,
-      color,
-    }),
-  });
+  const handleColor = async () =>
+    await fetch(`${BACKEND_URL}/user/changeColor`, {
+      method: 'POST',
+      headers,
+      credentials: 'include',
+      body: JSON.stringify({
+        id,
+        color,
+      }),
+    });
+  id && handleColor();
 };
